sorted out signs process into different packages and files
